+++ tests/unit/test_client.py
@@ -8,6 +8,7 @@ import pytest
 from spnkr.client import HaloInfiniteClient
 from spnkr.services import (
     DiscoveryUgcService,
+    EconomyService,
     GameCmsHacsService,
     ProfileService,
     SkillService,
@@ -38,6 +39,7 @@ def test_client_set_tokens(client: HaloInfiniteClient):
 def test_client_services(client: HaloInfiniteClient):
     """Test that the client services are created as expected."""
     assert isinstance(client.discovery_ugc, DiscoveryUgcService)
+    assert isinstance(client.economy, EconomyService)
     assert isinstance(client.gamecms_hacs, GameCmsHacsService)
     assert isinstance(client.profile, ProfileService)
     assert isinstance(client.skill, SkillService)